+++ src/Certificates/Commands/GetCertificateCommand.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Laravel\Forge\Certificates\Commands;
+
+class GetCertificateCommand extends CertificateCommand
+{
+}